% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.table.R
\name{as.dbi.table}
\alias{as.dbi.table}
\title{Coerce to a DBI Table}
\usage{
as.dbi.table(x, conn)
}
\arguments{
\item{x}{any \R object that can be coerced to a
\code{\link[base]{data.frame}}.}

\item{conn}{a connection handle returned by \code{\link[DBI]{dbConnect}}.
Alternatively, \code{conn} may be a \code{\link{dbi.table}} or a
\code{\link{dbi_database}}; in these cases, the connection handle
is extracted from the provided object.}
}
\description{
Write a \code{\link[base]{data.frame}} to a temporary table on a
             \code{DBI} connection and return a \code{\link{dbi.table}}.
}
\section{Note}{
 The temporary tables created by this function are dropped
               (by calling \code{\link[DBI]{dbRemoveTable}}) during garbage
               collection.
}

