% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csql}
\alias{csql}
\title{See SQL}
\usage{
csql(x, n = session$max_fetch, offset = NULL)
}
\arguments{
\item{x}{a \code{\link{dbi.table}}.}

\item{n}{a single integer value.}

\item{offset}{a single integer value or \code{NULL}.}
}
\description{
View a \code{\link{dbi.table}} as a SQL query.
}
