% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExecute.R
\docType{methods}
\name{dbExecute,dbi.catalog,ANY-method}
\alias{dbExecute,dbi.catalog,ANY-method}
\alias{dbExecute,dbi.schema,ANY-method}
\alias{dbExecute,dbi.table,ANY-method}
\title{Change database state}
\usage{
\S4method{dbExecute}{dbi.catalog,ANY}(conn, statement, ...)

\S4method{dbExecute}{dbi.schema,ANY}(conn, statement, ...)

\S4method{dbExecute}{dbi.table,ANY}(conn, statement, ...)
}
\arguments{
\item{conn}{a \code{\link{dbi.catalog}}, \code{dbi.schema}, or
\code{\link{dbi.table}}.}

\item{statement}{a character string containing SQL.}

\item{...}{other parameters passed on to methods.}
}
\value{
a scalar numeric that specifies the number of rows affected by the
        statement.
}
\description{
Execute a statement and returns the number of rows affected.
}
\details{
This function calls \code{\link[DBI]{dbExecute}} using the
         connection handle extracted from a \code{\link{dbi.catalog}},
         \code{dbi.schema}, or \code{\link{dbi.table}}.
}
\examples{
duck <- dbi.catalog(chinook.duckdb)
dev.null <- dbExecute(duck, "INSTALL httpfs;")
dev.null <- dbExecute(duck, "LOAD httpfs;")

}
\seealso{
\code{\link[DBI]{dbExecute}}
}
