% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_methods.R
\docType{methods}
\name{dbi.table.DBI}
\alias{dbi.table.DBI}
\alias{dbExecute,dbi.catalog,ANY-method}
\alias{dbExecute,dbi.schema,ANY-method}
\alias{dbExecute,dbi.table,ANY-method}
\alias{dbGetInfo,dbi.catalog-method}
\alias{dbGetInfo,dbi.catalog}
\alias{dbGetInfo,dbi.schema-method}
\alias{dbGetInfo,dbi.schema}
\alias{dbGetInfo,dbi.table-method}
\alias{dbGetInfo,dbi.table}
\alias{dbGetQuery,dbi.table,missing-method}
\alias{dbGetQuery,dbi.catalog,ANY-method}
\alias{dbGetQuery,dbi.schema,ANY-method}
\alias{dbGetQuery,dbi.table,ANY-method}
\alias{dbSendStatement,dbi.table,missing-method}
\alias{dbSendStatement,dbi.catalog,ANY-method}
\alias{dbSendStatement,dbi.schema,ANY-method}
\alias{dbSendStatement,dbi.table,ANY-method}
\title{DBI Methods for \code{dbi.table}s}
\usage{
\S4method{dbExecute}{dbi.catalog,ANY}(conn, statement, ...)

\S4method{dbExecute}{dbi.schema,ANY}(conn, statement, ...)

\S4method{dbExecute}{dbi.table,ANY}(conn, statement, ...)

\S4method{dbGetInfo}{dbi.catalog}(dbObj, ...)

\S4method{dbGetInfo}{dbi.schema}(dbObj, ...)

\S4method{dbGetInfo}{dbi.table}(dbObj, ...)

\S4method{dbGetQuery}{dbi.table,missing}(conn, statement, ...)

\S4method{dbGetQuery}{dbi.catalog,ANY}(conn, statement, ...)

\S4method{dbGetQuery}{dbi.schema,ANY}(conn, statement, ...)

\S4method{dbGetQuery}{dbi.table,ANY}(conn, statement, ...)

\S4method{dbSendStatement}{dbi.table,missing}(conn, statement, ...)

\S4method{dbSendStatement}{dbi.catalog,ANY}(conn, statement, ...)

\S4method{dbSendStatement}{dbi.schema,ANY}(conn, statement, ...)

\S4method{dbSendStatement}{dbi.table,ANY}(conn, statement, ...)
}
\arguments{
\item{conn}{a \code{\link{dbi.catalog}}, \code{dbi.schema}, or \code{\link{dbi.table}}.}

\item{statement}{a \code{\link[DBI]{SQL}} object.}

\item{\dots}{other parameters passed on to methods.}

\item{dbObj}{a \code{\link{dbi.catalog}}, \code{dbi.schema}, or \code{\link{dbi.table}}.}
}
\description{
Call DBI methods using the underlying DBI connection.
}
\seealso{
\code{\link[DBI]{dbExecute}}, \code{\link[DBI]{dbGetInfo}},
  \code{\link[DBI]{dbSendStatement}}
}
