% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.table.R
\name{dbi.table}
\alias{dbi.table}
\title{Create a \code{dbi.table} accessible via a \code{DBI} connection}
\usage{
dbi.table(conn, id)
}
\arguments{
\item{conn}{a connection handle returned by \code{\link[DBI]{dbConnect}}.}

\item{id}{an \code{\link[DBI]{Id}} or a character string identifying a
table accessible via a \code{conn}.}
}
\description{
Create a \code{dbi.table} from an existing SQL object.
}
