% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbi_database}
\alias{dbi_database}
\title{Create a \code{dbi_database}}
\usage{
dbi_database(conn)
}
\arguments{
\item{conn}{a connection handle returned by \code{\link[DBI]{dbConnect}} or
a zero-argument function that returns a connection handle.}
}
\value{
\code{dbi_database} returns a \code{dbi_database} (internally an
        \code{\link[base]{environment}} with the class attribute set to
        \code{"dbi_database"}).
}
\description{
Use \code{dbi_database} to create a \code{dbi_database},
\code{list_db_objects} to get a list of database objects (e.g., tables and
views) that can be added to the \code{dbi_database}, and
\code{add_db_objects} to add one or more of these objects. See Details and
Examples.
}
\details{
A \code{dbi_database} is an \code{environment} with the class
attribute set to \code{"dbi_database"}. Initially, the \code{dbi_database}
contains only the \code{information_schema}. When a database object
(e.g., a table or view identified by an \code{\link[DBI]{Id}}) is
added to the \code{dbi_database} using \code{add_db_objects}, a
\code{\link{dbi.table}} is created in the \code{dbi_database}. Since the
\code{dbi_database} is an envionment, the added \code{\link{dbi.table}} can
be accessed using the \code{$} operator. Further, one can use
\code{\link[base]{ls}} and \code{\link[base]{objects}} to list the
\code{\link{dbi.table}}s in the \code{dbi_database}. See Examples.
}
\examples{
# chinook.sqlite is a zero-argument function that returns a DBI handle
db <- dbi_database(chinook.duckdb)

# a dbi_database corresponds to a catalog - list the schemas
ls(db)

# list the tables in the schema 'main'
ls(db$main)

}
