% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{chinook.sqlite}
\alias{chinook.sqlite}
\alias{chinook.duckdb}
\title{Example Databases}
\usage{
chinook.sqlite()

chinook.duckdb()
}
\description{
These functions return connections to example databases
             included in the \pkg{dbi.table} package.
}
\section{Functions}{
\itemize{
\item \code{chinook.sqlite()}: 

\item \code{chinook.duckdb()}: 

}}
