% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_test.R
\name{reference_test}
\alias{reference_test}
\title{\code{dbi.table} test vs. reference implementation}
\usage{
reference_test(
  expr,
  envir = parent.frame(),
  ignore.row.order = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{expr}{an expression involving at least one \code{dbi.table}.}

\item{envir}{an environment. Where to evaluate \code{expr}.}

\item{ignore.row.order}{a logical value. This argument is passed to
\code{\link{all.equal}}.}

\item{verbose}{a logical value. When \code{TRUE}, the output from
\code{all.equal} is displaued in a message when
\code{all.equal} returns anything other than \code{TRUE}.}
}
\description{
Evaluate an expression using \code{dbi.table}s and separately using
\code{data.tables}s, then compare the results.
}
\details{
This function evaluates the input \code{x} in two different ways:

  1. (the typical dbi.table workflow) the input is translated to sql,
     evaluated by the connected database, and the results are downloaded as
     a data.table.

  2. (reference) the entire database table is downloaded to as data.table
     then the expression is evaluated locally.

The goal of the dbi.table package is to match the reference result. Matching
up to row order is acceptable.
}
