% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{as.dbi.table}
\alias{as.dbi.table}
\title{Coerce to a DBI Table}
\usage{
as.dbi.table(x, conn, row.names = FALSE)
}
\arguments{
\item{x}{an \R object coercable to a \code{\link[base]{data.frame}}.}

\item{conn}{a connection handle returned by \code{\link[DBI]{dbConnect}}.}

\item{row.names}{see \code{row.names} in \code{\link[DBI]{dbWriteTable}}.}
}
\description{
Write a \code{\link[base]{data.frame}} to temporary table and
             return a \code{\link{dbi.table}}.
}
