% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csql}
\alias{csql}
\title{See SQL}
\usage{
csql(x, n = getOption("dbi_table_max_fetch", 10000L))
}
\arguments{
\item{x}{a \code{dbi.table}.}

\item{n}{a single integer value. When nonnegative, limits the number of rows
returned by the query to \code{n}.}
}
\value{
none (invisible \code{NULL}).
}
\description{
View a \code{\link{dbi.table}}'s underlying SQL query.
}
