% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csql}
\alias{csql}
\title{See SQL}
\usage{
csql(x, n = session$max_fetch)
}
\arguments{
\item{x}{a \code{dbi.table}.}

\item{n}{a single integer value. When nonnegative, limits the number of rows
returned by the query to \code{n}.}
}
\description{
View a \code{\link{dbi.table}} as a SQL query.
}
