% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{dbi.attach}
\alias{dbi.attach}
\title{Attach a Database Schema to the Search Path}
\usage{
dbi.attach(what, pos = 2L, name = NULL, warn.conflicts = FALSE)
}
\arguments{
\item{what}{a connection handle returned by \code{\link[DBI]{dbConnect}} or
a zero-argument function that returns a connection handle.}

\item{pos}{an integer specifying position in \code{\link[base]{search}}()
where to attach.}

\item{name}{a character string specifying the name to use for the attached
database.}

\item{warn.conflicts}{a logical value. If \code{TRUE}, warnings are
printed about \code{\link[base]{conflicts}} from
attaching the database, unless that database contains
an object \code{.conflicts.OK}. A conflict is a
function masking a function, or a non-function
masking a non-function.}
}
\description{
Create a \code{\link{dbi.table}} for each database object in
             a schema and place them on the search path.
}
\seealso{
\code{\link[base]{attach}}
}
