% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{dbi.attach}
\alias{dbi.attach}
\title{Attach a DBI Connection to the Search Path}
\usage{
dbi.attach(
  what,
  pos = 2L,
  name = NULL,
  warn.conflicts = FALSE,
  ...,
  prefix = NULL
)
}
\arguments{
\item{what}{a connection handle returned by \code{\link[DBI]{dbConnect}}.}

\item{pos}{an integer specifying position in \code{\link[base]{search}}()
where to attach.}

\item{name}{a character string specifying the name to use for the attached
database.}

\item{warn.conflicts}{a logical value. If \code{TRUE}, warnings are
printed about \code{\link[base]{conflicts}} from
attaching the database, unless that database contains
an object \code{.conflicts.OK}. A conflict is a
function masking a function, or a non-function
masking a non-function.}

\item{\dots}{additional arguments are passed to
\code{\link[DBI]{dbListObjects}}.}

\item{prefix}{see the \code{prefix} argument in
\code{\link[DBI]{dbListObjects}}.}
}
\description{
Attach all (or a subset) of the database objects (tables, views,
             etc.) accessible over a DBI connection to the search path. The
             attached objects can be queried using
             \code{\link[data.table]{data.table}} syntax.
}
\note{
Attaches each object returned by \code{\link[DBI]{dbListObjects}}.
}
\seealso{
\code{\link[base]{attach}}
}
