% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ex_chinook}
\alias{ex_chinook}
\title{Example Databases}
\usage{
ex_chinook()
}
\description{
These functions return connections to example databases
             included in the \pkg{dbi.table} package.
}
\section{Functions}{
\itemize{
\item \code{ex_chinook()}: 

}}
