% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{chinook.sqlite}
\alias{chinook.sqlite}
\title{Example Databases}
\usage{
chinook.sqlite(type = c("Pool", "DBI"))
}
\arguments{
\item{type}{a character string. Use \code{"pool"} or \code{"DBI"}.}
}
\description{
These functions return connections to example databases
             included in the \pkg{dbi.table} package.
}
\section{Functions}{
\itemize{
\item \code{chinook.sqlite()}: 

}}
