% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{chinook.sqlite}
\alias{chinook.sqlite}
\title{Example Databases}
\usage{
chinook.sqlite(type = c("DBI", "path"))
}
\arguments{
\item{type}{a scalar character string. Use \code{"DBI"} for a
\code{DBIConnection} or \code{path} for the database file path.}
}
\description{
These functions return connections to example databases
             included in the \pkg{dbi.table} package.
}
\section{Functions}{
\itemize{
\item \code{chinook.sqlite()}: 

}}
