% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_databases}
\alias{example_databases}
\alias{chinook.sqlite}
\alias{chinook.duckdb}
\title{Example Databases}
\usage{
chinook.sqlite()

chinook.duckdb()
}
\value{
a \code{\link[DBI:DBIConnection-class]{DBIConnection}} object, as
  returned by \code{\link[DBI]{dbConnect}}.
}
\description{
These zero-argument functions return connections to the example
             databases included in the \pkg{dbi.table} package.
}
