% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_schema.R
\name{information_schema}
\alias{information_schema}
\title{Create an Information Schema for a \code{DBIConnection}}
\usage{
information_schema(conn)
}
\arguments{
\item{conn}{a connection handle returned by \code{\link[DBI]{dbConnect}}.
Note: unlike other functions in the \code{dbi.table} package,
\code{conn} must be the actual
\code{\link[DBI]{DBIConnection-class}} handle returned by
\code{\link[DBI]{dbConnect}} and NOT a function that creates one.}
}
\value{
an \code{\link[base]{environment}} containing the componenets
        enumerated in Details.
}
\description{
Caution: this function is not intended for end-users. Unless you are a
developer adding a method for an RDBMS that does not provide an Information
Schema, you do not need to work with this function directly; it is called by
\code{\link{dbi_database}} when initializing a new \code{dbi_database}.
}
\details{
An \code{information_schema} is an \code{\link[base]{environment}} that
contains
 \enumerate{
 \item a \code{\link[DBI]{DBIConnection-class}} named \code{.dbi_connection},
 \item a \code{\link[data.table]{data.table}} or \code{\link{dbi.table}}
       named \code{TABLES}, and
 \item a \code{\link[data.table]{data.table}} or \code{\link{dbi.table}}
       named \code{COLUMNS}.
}
The \code{TABLES} \code{*.table} must have columns named
\code{TABLE_CATALOG}, \code{TABLE_SCHEMA}, and \code{TABLE_NAME}. If
\code{TABLE_CATALOG} and/or \code{TABLE_SCHEMA} are not needed they may be
\code{NA_character_}. \code{TABLE_NAME} must be type \code{character} and
cannot have any missing values (\code{NA}s).

The \code{COLUMNS} \code{*.table} must have columns named
\code{TABLE_CATALOG}, \code{TABLE_SCHEMA}, \code{TABLE_NAME},
\code{COLUMN_NAME}, and \code{ORDINAL_POSITION}. If
\code{TABLE_CATALOG} and/or \code{TABLE_SCHEMA} are not needed they may be
\code{NA_character_}. \code{TABLE_NAME} and \code{COLUMN_NAME} must be type
\code{character} and cannot have any missing values (\code{NA}s).

\code{information_schema} is an S3 generic. The default method uses the
RDBMS's Information Schema if it exists (in which case \code{TABLES} and
\code{COLUMNS} are \code{\link{dbi.table}}s). Otherwise, it uses
\code{\link[DBI]{dbListTables}} and \code{\link[DBI]{dbListFields}} to create
a \emph{bare bones} information schema.
}
